requirements.txt
numpy==1.19.2
pandas
tensorflow
tensorflow_datasets
joblib
pickle-mixin
bert-for-tf2
sentencepiece
tensorflow_hub
bs4
bert
bert-tensorflow
lxml
sentencepiece
gast==0.4.0
grpcio<2.0,>=1.37.0
h5py~=3.1.0
tf-models-nightly
tf-nightly
tf-models-official==2.4.0
